requirements.txt
streamlit==1.26.0
langchain-community==0.0.227
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.1.0
tqdm
